+++ controllers/user.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/Keanus-In-Reevesverse/website-structure/database"
 	"github.com/Keanus-In-Reevesverse/website-structure/models"
+	"github.com/Keanus-In-Reevesverse/website-structure/services"
 	"github.com/gin-gonic/gin"
 )
 
@@ -19,9 +20,9 @@ func NewUser(c *gin.Context) {
 		return
 	}
 
-	//Creates
+	//Password encode && create
+	userCreated.Password = services.SHA256Encoder(userCreated.Password)
 	user := database.CreateUser(&userCreated)
-	database.CreateLogin(&userCreated)
 
 	//Response
 	var userReturn models.UserOutput
