+++ database/db.go
@@ -29,17 +29,9 @@ func CreateUser(userCreated *models.UserInput) *models.User {
 	user.Name = userCreated.Name
 	user.Email = userCreated.Email
 	user.PhoneNumber = userCreated.PhoneNumber
+	user.Password = userCreated.Password
 
 	DB.Table("USER").Create(&user)
 
 	return &user
-}
-
-func CreateLogin(userCreated *models.UserInput) {
-	var login models.Login
-
-	login.Email = userCreated.Email
-	login.Password = userCreated.Password
-
-	DB.Table("LOGIN").Create(&login)
 }
