+++ migration/docker-database-initial.sql
@@ -20,7 +20,8 @@ CREATE TABLE USER (
 user_id INT auto_increment PRIMARY KEY,
 email VARCHAR(60) UNIQUE,
 name VARCHAR(60),
-phone_number VARCHAR(12)
+phone_number VARCHAR(12),
+password VARCHAR(100)
 );
 
 CREATE TABLE FAVORITE (
@@ -32,14 +33,6 @@ FOREIGN KEY (game_id) REFERENCES GAME(game_id)
 );
 -- ALTER TABLE FAVORITE ADD FOREIGN KEY(user_id) REFERENCES USER (user_id)
 -- ALTER TABLE FAVORITE ADD FOREIGN KEY(game_id) REFERENCES GAME (game_id)
-
-CREATE TABLE LOGIN (
-user_id INT PRIMARY KEY,
-email VARCHAR(60) UNIQUE,
-password VARCHAR(30),
-FOREIGN KEY (user_id) REFERENCES USER(user_id)
-);
--- FOREIGN KEY(email,,) REFERENCES USER (user_id,email)
 
 CREATE TABLE PRICE_ALERT (
 user_id INT,
@@ -82,11 +75,9 @@ INSERT INTO GENRE VALUES (NULL, 'ACAO');
 
 INSERT INTO GAME VALUES (NULL, (SELECT GENRE_id FROM GENRE WHERE DESCRIPTION = 'ACAO'), 'FOR HONOR', 'UBSOFT', 'FOR HONOR É UM JOGO', NULL, 'LINK FDC');
 
-INSERT INTO USER VALUES(NULL, 'EMAIL@FATEC', 'NOME ALEATORIO', '11945473106');
+INSERT INTO USER VALUES(NULL, 'EMAIL@FATEC', 'NOME ALEATORIO', '11945473106', 'S3NH@_@L3@T0R1@');
 
 INSERT INTO FAVORITE VALUES((SELECT USER_id FROM USER WHERE EMAIL = 'EMAIL@FATEC'), (SELECT GAME_id FROM GAME WHERE NAME = 'FOR HONOR'));
-
-INSERT INTO LOGIN VALUES ( (SELECT USER_id FROM USER WHERE EMAIL = 'EMAIL@FATEC'), 'EMAIL@FATEC', 'Senha'); 
 
 INSERT INTO PRICE_ALERT VALUES ( (SELECT USER_id FROM USER WHERE EMAIL = 'EMAIL@FATEC'), (SELECT GAME_id FROM GAME WHERE NAME = 'FOR HONOR'), 15.50);
