+++ models/login.go
@@ -4,3 +4,16 @@ type Login struct {
 	Email    string
 	Password string
 }
+
+type UserInput struct {
+	Name        string
+	Email       string
+	Password    string
+	PhoneNumber string
+}
+
+type UserOutput struct {
+	Name        string
+	Email       string
+	PhoneNumber string
+}
