+++ models/user.go
@@ -7,20 +7,9 @@ type User struct {
 	Name        string
 	Email       string
 	PhoneNumber string
-}
-
-type UserInput struct {
-	Name        string
-	Email       string
 	Password    string
-	PhoneNumber string
 }
 
-type UserOutput struct {
-	Name        string
-	Email       string
-	PhoneNumber string
-}
 type History struct {
 	GameId     int64
 	StoreName  string
